abnormal-file-hub-main/backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<3.15
django-cors-headers>=4.3.0
gunicorn>=20.0,<21.0
python-dotenv>=1.0.0
whitenoise>=6.0,<7.0
pathspec==0.11.2
django-filter>=24.1 